move publish function into container + remove loadClassification use
